Migrate counting sort
- Move implementation to header file
- Add missing curly braces
- Add to sorting.cpp unit test file
- Modify unit test file to sort shorts instead of ints
- Update README
- Delete old counting_sort.cpp

Note on shorts instead of ints: The utility function that generates
unsorted vectors loads them with integers within a given range. Since
counting sort is glacially slow when there is a wide range between the
smallest and largest value, we're limiting this range from the minimum
value of a short to the maximum value.